Changed data type to GLfloat

+++ WcPt3D.h
@@ -20,16 +20,16 @@ class WcPt3D
 {
 public:
     WcPt3D() : _x(0.0), _y(0.0), _z(0.0) {}
-    WcPt3D(GLdouble x, GLdouble y, GLdouble z) : _x(x), _y(y), _z(z) {}
+    WcPt3D(GLfloat x, GLfloat y, GLfloat z) : _x(x), _y(y), _z(z) {}
 
-    GLdouble getX() const { return _x; }
-    GLdouble getY() const { return _y; }
-    GLdouble getZ() const { return _z; }
+    GLfloat getX() const { return _x; }
+    GLfloat getY() const { return _y; }
+    GLfloat getZ() const { return _z; }
 
-    void setX(GLdouble x) { _x = x; }
-    void setY(GLdouble y) { _y = y; }
-    void setZ(GLdouble z) { _z = z; }
+    void setX(GLfloat x) { _x = x; }
+    void setY(GLfloat y) { _y = y; }
+    void setZ(GLfloat z) { _z = z; }
 
 protected:
-    GLdouble _x, _y, _z;
+    GLfloat _x, _y, _z;
 };
